# Alternative fallback fields implementation