--- a/crates/rune-macros/src/any.rs
+++ b/crates/rune-macros/src/any.rs
@@ -665,7 +665,12 @@ where
             };
 
             #[cfg(not(feature = "dynamic_fields"))]
-            let meta_fields = quote! {};
+            let meta_fields = quote! {
+                #[automatically_derived]
+                impl #impl_generics #dynamic_field_search for #ident #type_generics #where_clause {
+                    const DYNAMIC_FIELD_MODE: #dynamic_field_mode = #dynamic_field_mode::Never;
+                }
+            };
 
             Some(quote! {
                 #[automatically_derived]