--- a/crates/rune/src/runtime/value.rs
+++ b/crates/rune/src/runtime/value.rs
@@ -1,5 +1,4 @@
-mod serde;
-
+use ::rust_alloc::sync::Arc;
 use core::any;
 use core::borrow::Borrow;
 use core::cmp::{Eq, Ord, Ordering, PartialEq, PartialOrd};
@@ -7,11 +6,13 @@ use core::fmt;
 use core::hash;
 use core::ptr;
 
-use ::rust_alloc::sync::Arc;
+use ::serde::{Deserialize, Serialize};
 
 use crate::alloc::fmt::TryWrite;
 use crate::alloc::prelude::*;
 use crate::alloc::{self, String};
+#[cfg(feature = "dynamic_fields")]
+use crate::compile::dynamic_fields::DynamicFieldMode;
 use crate::compile::ItemBuf;
 use crate::runtime::vm::CallResult;
 use crate::runtime::{
@@ -25,7 +26,7 @@ use crate::runtime::{
 use crate::runtime::{Hasher, Tuple};
 use crate::{Any, Hash};
 
-use ::serde::{Deserialize, Serialize};
+mod serde;
 
 // Small helper function to build errors.
 fn err<T, E>(error: E) -> VmResult<T>
@@ -1210,6 +1211,17 @@ impl Value {
             Self::Variant(empty) => vm_try!(empty.borrow_ref()).type_info(),
             Self::Any(any) => vm_try!(any.borrow_ref()).type_info(),
         })
+    }
+
+    #[cfg(feature = "dynamic_fields")]
+    /// Function returns the method in which to search for fields within
+    /// the value.
+    pub(crate) fn field_mode(&self) -> Result<DynamicFieldMode, VmError> {
+        if let Value::Any(v) = self {
+            Ok(v.borrow_ref()?.field_mode())
+        } else {
+            Ok(DynamicFieldMode::Never)
+        }
     }
 
     /// Perform a partial equality test between two values.