… crates/rune-macros/Cargo.toml
… crates/rune/Cargo.toml
… crates/rune/src/runtime/vm.rs
… crates/rune/src/tests.rs
… crates/rune/src/tests/bug_344.rs
… crates/rune/src/tests/dynamic_fields.rs