@udoprog 2022-10-23
I'm gonna have to think this over. It's not a style choice that I would choose, so if I include it I'd like to see the following changes done:
* Introduce your feature as extra Vm instructions which are feature gated. So instead of rewriting the `ObjectSlotIndexGet` instruction, add a new one which uses the new field loading convention.
* Feature gate the extra plumbing in `AnyObj` which allows for dynamic index get conventions.
* Add a [compiler option](https://docs.rs/rune/latest/rune/compile/struct.Options.html) which changes how index gets are performed so that they make use of the new feature. This would allow you to use the new feature, but it wouldn't be enabled by default.

That way the feature can be tested and refined in parallel.

#### Concerns

1) Dynamically deciding on the field loading convention in use is a form of metaprogramming which doesn't obviously mesh well with a future gradual type system. 2) Having dynamic field loading conventions might make future compile time optimizations harder or not possible.

Both of these are *future potentials* and not current blockers, which is why I wouldn't mind including the feature on an experimental basis to get a feel for it. Maybe improve it and mainline include in the future.

Thank you!