@LoopyAshy 2022-10-23
> I'm gonna have to think this over. It's not a style choice that I would choose, so if I include it I'd like to see the following changes done:
> 
> * Introduce your feature as extra Vm instructions which are feature gated. So instead of rewriting the `ObjectSlotIndexGet` instruction, add a new one which uses the new field loading convention.
> * Feature gate the extra plumbing in `AnyObj` which allows for dynamic index get conventions.
> * Add a [compiler option](https://docs.rs/rune/latest/rune/compile/struct.Options.html) which changes how index gets are performed so that they make use of the new feature. This would allow you to use the new feature, but it wouldn't be enabled by default.
> 
> That way the feature can be tested and refined in parallel.
> 
> #### Concerns
> 1. Dynamically deciding on the field loading convention in use is a form of metaprogramming which doesn't obviously mesh well with a future gradual type system. 2) Having dynamic field loading conventions might make future compile time optimizations harder or not possible.
> 
> Both of these are _future potentials_ and not current blockers, which is why I wouldn't mind including the feature on an experimental basis to get a feel for it. Maybe improve it and mainline include in the future.
> 
> Thank you!

Thank you for your reply. I am happy to make those changes and will do so when I am free.

Regarding you concerns: I do believe it shouldn't make it completely impossible but I can see the possibility of increasing the complexity of the compilation, if it does come to that I am however more than happy to try and assist in the ease of implementation as this is a QoL feature that I know will be very useful in quite a few of my projects for example in one of my projects the types are generated at load time and stored as templates and generated dynamically at runtime when requested and are optimised around these facts.
Reason for this is so the project doesn't need recompiled as it relies on third party data types which are very often change but the layouts are available, however there layouts are statically typed so if I gave a float instead of a int it would be incorrect, so I have to check the typing and field name validity etc. 

When I was using Lua for this project I simply overwrote the meta function for getting and setting fields and it worked like any other type and looked no different, it was seamless and effective. 
When moving to rune due to it having many many preferred design decisions over lua, it was only possible by having dedicated instance functions or using INDEX_* which means using `["field_name"]` which makes it look strange to the script writers who think of these types are true types not `HashMaps` or collections.