@udoprog 2022-10-23
> The AnyObj runtime support (just nice to avoid the vtable increase if it's not used but not super important).

Actually, no this can't be done. `AnyObjVtable` is public API! So it would suffer the same issues as the `Any` trait. The field needs to be unconditionally included.