@LoopyAshy 2022-10-24
I'll add the requested changes when I get back later as I am busy today, thank you for the review <3