@udoprog 2023-11-01
Naming can always be punted on, initially this will be released as a `--cfg` flag, because enabling it as a feature could end up changing the behavior of the project in unrelated places. Features either have to be adopted wholesale and define the semantics of the language or optionally opted into like that.

That being said, after digging into this now I'm prone to cutting down this feature by only supporting `DYNAMIC_FIELD_GET` / `DYNAMIC_FIELD_SET`. If any of these protocols are registered it indicates that the implementer takes over full control of field getting and setting.

I find the semantics as it currently stands with the different modes to be more confusing to describe than such a method, and both would ultimately support the same set of use cases.
E.g.
If `first` is set and DYNAMIC_FIELD_GET returns `None` then it falls back to `GET`.

There's also a lack of reflexivity in the features. Only `DYNAMIC_FIELD_GET` supports falling back to default `GET`, `DYNAMIC_FIELD_SET` does not because that would involve some kind of sentinel value indicating whether the field was set or not.

So this is an instance where I think it's better to define the behavior in code, rather than metadata since it both leads to clearer semantics, and a smaller change. To this end I've pushed a suggestion I'd like you to take for a spin.